int & int_vec

+++ src/threads/utils.cpp
@@ -10,7 +10,7 @@ extern "C" float _tid(Scope_Struct *scope_struct) {
 
 int last_thread_id=1;
 
-extern "C" float tid(Scope_Struct *scope_struct) {
-    float thread_id = scope_struct->thread_id-1;
+extern "C" int tid(Scope_Struct *scope_struct) {
+    int thread_id = scope_struct->thread_id-1;
     return thread_id;
 }
